+++ src/modules/loyalty/LoyaltyRoot.sol
@@ -10,15 +10,15 @@ import {PointManager} from "./PointManager.sol";
  * @dev Hợp đồng chính (root) cho module Loyalty.
  */
 contract LoyaltyRoot is Initializable {
-    uint256 public branchId;
-    uint256 public orgId;
+    uint48 public branchId;
+    uint48 public orgId;
     address public staffManager;
     address public pointManager;
     address public branchModuleManager;
 
     function initialize(
-        uint256 _branchId,
-        uint256 _orgId,
+        uint48 _branchId,
+        uint48 _orgId,
         address _staffManager,
         address _pointManagerBeacon,
         address _branchModuleManager
@@ -28,10 +28,15 @@ contract LoyaltyRoot is Initializable {
         staffManager = _staffManager;
         branchModuleManager = _branchModuleManager;
 
-        pointManager = address(new BeaconProxy(
-            _pointManagerBeacon,
-            abi.encodeCall(PointManager.initialize, (_branchId, _orgId, _branchModuleManager))
-        ));
+        pointManager = address(
+            new BeaconProxy(
+                _pointManagerBeacon,
+                abi.encodeCall(
+                    PointManager.initialize,
+                    (_branchId, _orgId, _branchModuleManager)
+                )
+            )
+        );
     }
 
     function getSubContracts() external view returns (address _pointManager) {
