+++ src/modules/loyalty/PointManager.sol
@@ -12,14 +12,17 @@ contract PointManager is BranchContextUpgradeable {
     mapping(address => uint256) public userPoints;
 
     function initialize(
-        uint256 _branchId,
-        uint256 _orgId,
+        uint48 _branchId,
+        uint48 _orgId,
         address _branchModuleManager
     ) external initializer {
         __BranchContext_init(_branchId, _orgId, _branchModuleManager);
     }
 
-    function addPoints(address user, uint256 amount) external onlyIfModuleEnabled(ModuleKeys.MODULE_LOYALTY) {
+    function addPoints(
+        address user,
+        uint256 amount
+    ) external onlyIfModuleEnabled(ModuleKeys.MODULE_LOYALTY) {
         userPoints[user] += amount;
     }
 }
