+++ src/modules/meos/AccountManager.sol
@@ -12,14 +12,17 @@ contract AccountManager is BranchContextUpgradeable {
     mapping(string => address) public usernameToWallet;
 
     function initialize(
-        uint256 _branchId,
-        uint256 _orgId,
+        uint48 _branchId,
+        uint48 _orgId,
         address _branchModuleManager
     ) external initializer {
         __BranchContext_init(_branchId, _orgId, _branchModuleManager);
     }
 
-    function registerUser(string calldata username, address wallet) external onlyIfModuleEnabled(ModuleKeys.MODULE_MEOS) {
+    function registerUser(
+        string calldata username,
+        address wallet
+    ) external onlyIfModuleEnabled(ModuleKeys.MODULE_MEOS) {
         usernameToWallet[username] = wallet;
     }
 }
